pyproject.toml
[project]
name = "thingking"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.4.0",
    "google-generativeai>=0.8.6",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.1.12",
    "langgraph>=0.6.11",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
